integration tests: move ethernet/common integration tests to a separate file
